Перенёс метод алерта в ВьюКонтроллер
