+++ MovieQuiz/Presentation/AlertPresenter.swift
@@ -1,25 +1,25 @@
-import UIKit
-
-final class AlertPresenter: AlertPresenterProtoсol {
-    
-    private weak var viewController: UIViewController?
-    
-    init(viewController: UIViewController) {
-        self.viewController = viewController
-    }
-    
-    func showAlert(in model: AlertModel) {
-        let alert = UIAlertController(
-            title: model.title,
-            message: model.message,
-            preferredStyle: .alert)
-        
-        alert.view.accessibilityIdentifier = "Game results"
-        
-        let action = UIAlertAction(title: model.buttonText, style: .default) { _ in
-            model.buttonAction()
-        }
-        alert.addAction(action)
-        viewController?.present(alert, animated: true)
-    }
-}
+//import UIKit
+//
+//final class AlertPresenter: AlertPresenterProtoсol {
+//    
+//    private weak var viewController: UIViewController?
+//    
+//    init(viewController: UIViewController) {
+//        self.viewController = viewController
+//    }
+//    
+//    func showAlert(in model: AlertModel) {
+//        let alert = UIAlertController(
+//            title: model.title,
+//            message: model.message,
+//            preferredStyle: .alert)
+//        
+//        alert.view.accessibilityIdentifier = "Game results"
+//        
+//        let action = UIAlertAction(title: model.buttonText, style: .default) { _ in
+//            model.buttonAction()
+//        }
+//        alert.addAction(action)
+//        viewController?.present(alert, animated: true)
+//    }
+//}
