+++ MovieQuiz/Services/AlertPresenterProtoсol.swift
@@ -1,5 +1,5 @@
-import Foundation
-
-protocol AlertPresenterProtoсol: AnyObject {
-    func showAlert(in model: AlertModel)
-    }
+//import Foundation
+//
+//protocol AlertPresenterProtoсol: AnyObject {
+//    func showAlert(in model: AlertModel)
+//    }
