+++ MovieQuizTests/MovieQuizPresenterTests.swift
@@ -3,8 +3,9 @@ import XCTest
 @testable import MovieQuiz
 
 final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
+    func showAlert(in: MovieQuiz.AlertModel) {}
     func isEnabledButton(_: Bool) {}
-    var alertPresenter: MovieQuiz.AlertPresenterProtoсol?
+//    var alertPresenter: MovieQuiz.AlertPresenterProtoсol?
     func show(quiz step: QuizStepViewModel) {}
     func highlightImageBorder(isCorrectAnswer: Bool) {}
     func showLoadingIndicator() {}
